Move TypeScript to comments
