+++ types/source-maps.ts
@@ -1,0 +1,10 @@
+export declare module LHSourceMap {
+  export interface Entry {
+    lineNumber: number;
+    columnNumber: number;
+    sourceURL: string;
+    sourceLineNumber: number;
+    sourceColumnNumber: number;
+    name: string;
+  }
+}
